fixed docs and moved squad_eval to tools
