+++ allennlp/models/reading_comprehension/dialog_qa.py
@@ -6,7 +6,7 @@ import torch
 import torch.nn.functional as F
 from torch.nn.functional import nll_loss
 
-from allennlp.common import squad_eval
+from allennlp.tools import squad_eval
 from allennlp.data import Vocabulary
 from allennlp.models.model import Model
 from allennlp.modules import Seq2SeqEncoder, TimeDistributed, TextFieldEmbedder
