+++ doc/api/allennlp.tools.rst
@@ -1,0 +1,15 @@
+allennlp.tools
+==============
+
+Modules containing official evaluators of various
+tasks for which we build models.
+
+.. toctree::
+
+   allennlp.tools.squad_eval
+   allennlp.tools.wikitables_evaluator
+
+.. automodule:: allennlp.tools
+   :members:
+   :undoc-members:
+   :show-inheritance:
