+++ doc/api/allennlp.tools.squad_eval.rst
@@ -1,7 +1,7 @@
-allennlp.common.squad_eval
+allennlp.tools.squad_eval
 ===============================
 
-.. automodule:: allennlp.common.squad_eval
+.. automodule:: allennlp.tools.squad_eval
    :members:
    :undoc-members:
    :show-inheritance:
