+++ doc/api/allennlp.tools.wikitables_evaluator.rst
@@ -1,0 +1,7 @@
+allennlp.tools.wikitables_evaluator
+===================================
+
+.. automodule:: allennlp.tools.wikitables_evaluator
+   :members:
+   :undoc-members:
+   :show-inheritance:
