+++ doc/index.rst
@@ -22,6 +22,7 @@ AllenNLP is a libary built on top of PyTorch to make NLP easier.
    api/allennlp.semparse
    api/allennlp.service
    api/allennlp.state_machines
+   api/allennlp.tools
    api/allennlp.training
    api/allennlp.pretrained
